pyproject.toml
[project]
name = "projeto2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.78.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
